[docs] Share source-safe compatibility validation
